backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis>=4.5.2,<5.0.0
celery==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Text Processing & Document Parsing
PyPDF2==3.0.1
pdfplumber==0.10.3
ebooklib==0.18
beautifulsoup4==4.12.2
nltk==3.8.1

# AI & TTS
TTS==0.21.1  # Coqui TTS
google-generativeai==0.3.2  # Gemini API
torch==2.1.1
torchaudio==2.1.1

# Audio Processing
pydub==0.25.1
librosa==0.10.1

# HTTP Requests & APIs
httpx==0.25.2
aiofiles==23.2.1

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Background Tasks
celery[redis]==5.3.4
flower==2.0.1

# File Handling
pillow==10.1.0

# Utilities
python-slugify==8.0.1
uuid==1.30

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring
structlog==23.2.0 